requirements.txt
ib_insync
pandas
numpy
aiosqlite
requests
loguru
fastapi
uvicorn
pyarrow
schedule
python-dotenv
rich
matplotlib
twilio
pytest
